new test file set up for newWorld.js testing
