--- functions/test/sampleTest.js
@@ -1,15 +1,0 @@
-// const assert = require('chai').assert;
-// const app = require('../app');
-// // const manage3dworlds = require('../public/js/manage3DWorld');
-//
-// describe('App', function(){
-// 	it('app should return hello', function(){
-// 		assert.equal();
-// 	});
-// });
-//
-// describe('manage3DWorld', function(){
-// 	it('should return hello', function(){
-// 		assert.equal();
-// 	});
-// });
